requirements.txt

Django==4.0
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0

